starting with ELM striation finder.
This program shall be able to plot the temporal heat flux evolution, plot specific timepoints in another frame.
Further it shall find ELMs in a given range between start and end time points specified by the user.
If ELMs are found the user is able to jump to the next ELM onset.