Changing position of striation by mouse-click.
The position mark by left-click now is the maximum in a given interval.
The interval is pre-set to a fixed value (5) but can be changed in the edit menu.

With shift+left-click the separatrix position can be marked.
This draws a line on the pposition of the cursor.
This needs some fixing, at the moment it has to be clicked at first after one point/striation.

The three save options are now ordered in a sub-menu.